# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User Management",
    "description": "Explore User Management APIs",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new User.",
        "operationId": "User.create",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Model instance data",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "deprecated": false
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "login API.",
        "operationId": "User.login",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Model instance data",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "deprecated": false
      }
    },
    "/users/getusers": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all Users API.",
        "operationId": "User.getusers",
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "deprecated": false
      }
    }
  },
  "tags": [
    {
      "name": "User"
    }
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "application/xml",
    "text/xml"
  ],
  "produces": [
    "application/json"
  ],
  "definitions": {
    "User": {
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "email",
        "password"
      ],
      "additionalProperties": false
    },
    "Login": {
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ],
      "additionalProperties": false
    },
    "x-any": {
      "properties": {}
    }
  }
}